Implemented A* algorithm for path finding. Works very well! Also, more work on GUI, and abstracted the controller out of the player so I can use the player class for computer controllers in the future.

+++ src/simplerts/display/Camera.java
@@ -24,7 +24,7 @@ public class Camera {
         offsetY = 0;
     }
     
-    public void addOffset(int x, int y){
+    public void addOffset(float x, float y){
         offsetX += x;
         offsetY += y;
